security: CodeRabbit 리뷰 3차 반영 — 미반영 이슈 전체 수정
백엔드:
- [Issue 4] 리프레시 토큰에 totp_verified 클레임 추가 → /refresh 경유 2FA 우회 차단
- [Issue 4] /2fa/verify 성공 시 refresh_token 로테이션 (totp_verified=True로 갱신)
- [Issue 5] /2fa/verify에 계정 잠금 체크 추가 (잠긴 계정으로 OTP 시도 차단)
- [Issue 10] /refresh에서 DB 세션 만료 시각 검증 추가 + expires_at 갱신
- [Issue 12] 2FA 재설정 시 pending_totp_secret 임시 저장 → 기존 활성 2FA 보호
- [Issue 13/14] record_login_failure: 만료 잠금 시 카운터 초기화 + 원자적 UPDATE
- [Issue 15] create_refresh_token에 jti 클레임 추가 (토큰 고유 ID)
- [Issue 16] get_or_create_user: IntegrityError 처리로 동시성 경쟁 조건 수정
- [Issue 17] crypto_service.py 문서 수정 (Fernet=AES-128, AES-256 마이그레이션 TODO)
- [Issue 18] /sessions is_current: JWT sid 클레임으로 현재 세션 정확히 표시
- [Issue 19] ExchangeRate.rate: Mapped[float] → Mapped[Decimal] (정밀도 손실 방지)
- [Issue 20] Currency enum에 ETH 추가 (프론트엔드 타입과 동기화)
- [Issue 26] Dockerfile: 비루트 사용자(appuser)로 실행
- [Issue 29] conftest.py: LifespanManager 추가 (DB lifespan 이벤트 실행)
- [Issue 35] asgi-lifespan requirements-dev.txt에 추가

DB 마이그레이션:
- asset_snapshots (user_id, snapshot_date) 유니크 제약 추가
- users.pending_totp_secret 컬럼 추가
- currency ENUM에 ETH 추가

프론트엔드:
- [Issue 21] useIdleTimeout: handleStorage에서 잔여 시간 계산 (늦은 이벤트 처리 대응)
- [Issue 23] api.ts: FormData 요청 시 Content-Type 미강제 (multipart 자동 설정 보호)
- [Issue 24] TotpVerifyResponse: access_token, refresh_token 필드 추가
- [Issue 32] login/page.tsx: decodeURIComponent 이중 호출 제거 (URIError 방지)
- [Issue 33] Header.tsx: 마스킹/테마 토글 버튼에 aria-label 추가 (접근성)

+++ backend/app/models/exchange_rate.py
@@ -1,6 +1,7 @@
 """환율 모델"""
 
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, Integer, Numeric, String, func
 from sqlalchemy.orm import Mapped, mapped_column
@@ -14,7 +15,7 @@ class ExchangeRate(Base):
     id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     from_currency: Mapped[str] = mapped_column(String(10), nullable=False, index=True)
     to_currency: Mapped[str] = mapped_column(String(10), nullable=False, index=True)
-    rate: Mapped[float] = mapped_column(Numeric(20, 8), nullable=False)
+    rate: Mapped[Decimal] = mapped_column(Numeric(20, 8), nullable=False)  # float 대신 Decimal (부동소수점 정밀도 손실 방지)
     fetched_at: Mapped[datetime] = mapped_column(
         DateTime(timezone=True), server_default=func.now(), nullable=False
     )
